Set up gradle with dependencies
